Rename permissions.js to perms.js
